+++ apps/dashboard/src/app/(dashboard)/(chain)/[chain_id]/[contractAddress]/_components/claim-conditions/claim-conditions-form/hooks.ts
@@ -15,7 +15,7 @@ import type { z } from "zod";
 import type {
   ClaimCondition as LegacyClaimCondition,
   ClaimConditionInputSchema as LegacyClaimConditionInputSchema,
-} from "../../legacy-zod-schema";
+} from "../legacy-zod-schema";
 
 type SnapshotEntry = {
   address: string;
